Update fi.csv
Expanding fi.csv with several Finnish web sites. Lines 7 to 10 contain material that is seen as politically or culturally inappropriate, are hosting media content and/or opinions that have led to police investigations and criminal sentences before and are prone to publish such material in the future. Because of this they are prone to be closed or blocked. But because site blocking is rare in Finland, such action would reflect somekind of political or juidical shift.
Lines 11 to 12 are related to Sami, the indigenous ethnic minority group living in nordic countries and Russia. Their pursuit for their rights are seen as a politically hot topic by some. If these sites are blocked or banned, it should be noted.